OS X Framework + Client (experimental)
A new OS X client app bundle and a Framework, using most of the
original code. This will serve as a reference for a new client made
from scratch using the API.

The Framework target in the XCode cannot be generated by cmake.
Normally it would be ignored but I committed it in this branch for
archiving purposes.